add tests; rename util.py to utils.py
