+++ gwcs/tests/data/simple_wcs2.hdr
@@ -1,0 +1,1 @@
+WCSAXES =                    2 / Number of coordinate axes                      CRPIX1  =                507.0 / Pixel coordinate of reference point            CRPIX2  =                507.0 / Pixel coordinate of reference point            PC1_1   =    7.70605644414E-06 / Coordinate transformation matrix element       PC1_2   =    3.29130820267E-05 / Coordinate transformation matrix element       PC2_1   =    3.68234230443E-05 / Coordinate transformation matrix element       PC2_2   =   -6.77287573742E-06 / Coordinate transformation matrix element       CDELT1  =                  1.0 / [deg] Coordinate increment at reference point  CDELT2  =                  1.0 / [deg] Coordinate increment at reference point  CUNIT1  = 'deg'                / Units of coordinate increment and value        CUNIT2  = 'deg'                / Units of coordinate increment and value        CTYPE1  = 'RA---TAN-SIP'       / Right ascension, gnomonic projection           CTYPE2  = 'DEC--TAN-SIP'       / Declination, gnomonic projection               CRVAL1  =        251.204239952 / [deg] Coordinate value at reference point      CRVAL2  =        57.5817453704 / [deg] Coordinate value at reference point      LONPOLE =                180.0 / [deg] Native longitude of celestial pole       LATPOLE =        57.5817453704 / [deg] Native latitude of celestial pole        WCSNAME = 'IDC_w3m18525i'      / Coordinate system title                        RADESYS = 'ICRS'               / Equatorial coordinate system                   END
